requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
weasyprint==57.0
google-generative-ai==0.9.0
Werkzeug==2.3.6
gunicorn==21.2.0